refactor: ExcludedRuleCount/ExcludedSummaryをexclude.rsへ移す (#28)
mod.rsが既に1249行のgod-moduleで、本PRで1404行まで増やしていた。
除外サマリの型はexclude.rsが生成するデータであり凝集度が高いため、
定義をそちらへ移し、mod.rsからはpub(crate) useで再輸出する。
外部から見える型パスとserdeのwire契約は変えない。
